restructure test utilities
