+++ test-app/tests/helpers/common.ts
@@ -1,0 +1,8 @@
+/**
+ * Promise which waits for ms milliseconds
+ * @param ms number of milliseconds to wait
+ * @returns A Promise which waits for ms milliseconds
+ */
+export function delayMs(ms: number): Promise<void> {
+  return new Promise((r) => setTimeout(r, ms));
+}
