+++ test-app/tests/unit/prosemirror/view-test.ts
@@ -1,11 +1,11 @@
 import { module, test } from 'qunit';
-import TEST_SCHEMA from '../../../tests/test-utils';
 import { EditorState, SayView } from '@lblod/ember-rdfa-editor';
 import { oneLineTrim } from 'common-tags';
+import { SAMPLE_SCHEMA } from 'test-app/tests/helpers/prosemirror';
 
 module('ProseMirror | view', function () {
   test('setHtmlContent without a supplied range should replace the whole content of the document', function (assert) {
-    const schema = TEST_SCHEMA;
+    const schema = SAMPLE_SCHEMA;
 
     const view = new SayView(null, {
       state: EditorState.create({ schema }),
@@ -61,7 +61,7 @@ module('ProseMirror | view', function () {
     assert.strictEqual(view.htmlContent, expectedHtml);
   });
   test('setHtmlContent should be able to replace a specific range when specified', function (assert) {
-    const schema = TEST_SCHEMA;
+    const schema = SAMPLE_SCHEMA;
 
     const view = new SayView(null, {
       state: EditorState.create({ schema }),
@@ -123,7 +123,7 @@ module('ProseMirror | view', function () {
   });
 
   test('setHtmlContent will leave not clean the HTML if asked not to', function (assert) {
-    const schema = TEST_SCHEMA;
+    const schema = SAMPLE_SCHEMA;
 
     const view = new SayView(null, {
       state: EditorState.create({ schema }),
